apps/orchestrator/pyproject.toml
[project]
name = "pulse-orchestrator"
version = "0.1.0"
description = "PULSE ADK multi-agent orchestrator. Runs the 5s decision loop, delegates to Flow / Queue / Concierge specialists, and writes interventions + cost-tracked traces back to Firestore."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "google-adk>=1.2.0",
  "google-genai>=1.7.0",
  "google-cloud-firestore>=2.19.0",
  "google-cloud-pubsub>=2.26.0",
  "pydantic>=2.8.0",
]

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
